Refactor rest-testkit #24
