+++ rest-testkit/src/test/scala/busymachines/rest_test/AuthenticatedRoutesTest.scala
@@ -2,6 +2,7 @@ package busymachines.rest_test
 
 import busymachines.rest._
 import busymachines.rest_test.routes._
+import org.scalatest.Outcome
 
 /**
   *
@@ -9,9 +10,18 @@ import busymachines.rest_test.routes._
   * @since 07 Sep 2017
   *
   */
-private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSupport {
-  private lazy val authAPI = new AuthenticatedRoutesRestAPIForTesting()
-  override implicit val testedRoute: Route = Route.seal(authAPI.route)
+private[rest_test] class AuthenticatedRoutesTest extends ExampleRestAPITestBaseClassWithFixture {
+
+  /**
+    * A fixture would have to be more complicated than this to warrant all the hassle
+    *
+    */
+  override protected def withFixture(test: OneArgTest): Outcome = {
+    val authAPI = new AuthenticatedRoutesRestAPIForTesting()
+    val r: RestAPI = RestAPI.seal(authAPI)
+    this._testedRoute = r.route
+    test(r)
+  }
 
   import busymachines.json.auto._
 
@@ -21,7 +31,7 @@ private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSu
 
   //===========================================================================
 
-  it should "... return 401 Unauthorized when trying to access route without authe" in {
+  it should "... return 401 Unauthorized when trying to access route without authe" in { _ =>
     context(Contexts.none) { implicit cc =>
       get("/authentication") {
         expectStatus(StatusCodes.Unauthorized)
@@ -31,7 +41,7 @@ private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSu
 
   //===========================================================================
 
-  it should "... return 200 OK when providing proper Basic authentication" in {
+  it should "... return 200 OK when providing proper Basic authentication" in { _ =>
     context(BasicAuthenticationContextForTesting) { implicit cc =>
       get("/authentication") {
         expectStatus(StatusCodes.OK)
@@ -46,7 +56,7 @@ private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSu
 
   //===========================================================================
 
-  it should "... return 200 OK when trying to access API with optional auth, while not providing it" in {
+  it should "... return 200 OK when trying to access API with optional auth, while not providing it" in { _ =>
     context(Contexts.none) { implicit cc =>
       get("/opt_authentication") {
         expectStatus(StatusCodes.OK)
@@ -61,7 +71,7 @@ private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSu
 
   //===========================================================================
 
-  it should "... return 200 OK when trying to access API with optional auth, while providing it" in {
+  it should "... return 200 OK when trying to access API with optional auth, while providing it" in { _ =>
     context(BasicAuthenticationContextForTesting) { implicit cc =>
       get("/opt_authentication") {
         expectStatus(StatusCodes.OK)
@@ -75,5 +85,4 @@ private[rest_test] class AuthenticatedRoutesTest extends RestAPITest with JsonSu
   }
 
   //===========================================================================
-
 }
